// src/app/shared/model/userDetail.model.ts
export class UserDetail {
  id: any;
  userName: string;
  password: string;
  emailID: string;
  imagePath: string;

  // constructor(
  //   id: string,
  //   userName: string,
  //   password: string,
  //   emailID: string,
  //   imagePath: string
  // ) {
  //   this.id = id;
  //   this.userName = userName;
  //   this.password = password;
  //   this.emailID = emailID;
  //   this.imagePath = imagePath;
  // }
}